+++ CHANGELOG.md
@@ -7,6 +7,20 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+#### Added
+
+- warning if the current scarb version doesn't support not including macros in the coverage report
+- `--unstable` flag to enable unstable features. Read more about
+  it [here](./README.md#coverage-across-different-scarb-versions)
+
+#### Changed
+
+- macros are now by default included in the coverage report. If you want to exclude them, use the `--include` without
+  the `macros` option (can also have empty value). This is due to a new stability policy. Read more about
+  it [here](./README.md#coverage-across-different-scarb-versions)
+- by default, the hit count of the lines will be truncated to 1. This can be changed with the `--no-truncation` flag.
+  This is due to a new stability policy. Read more about it [here](./README.md#coverage-across-different-scarb-versions)
+
 ## [0.4.0] - 2025-01-03
 
 #### Added